feat: implement usecase pattern for reward batch controller
