+++ src/test/java/it/gov/pagopa/idpay/transactions/service/RewardBatchServiceImplTest.java
@@ -1,67 +1,19 @@
 package it.gov.pagopa.idpay.transactions.service;
 
-import com.azure.core.http.rest.Response;
-import com.azure.storage.blob.models.BlockBlobItem;
-import com.mongodb.client.result.DeleteResult;
-import it.gov.pagopa.common.web.dto.ErrorDTO;
-import it.gov.pagopa.common.web.exception.*;
-import it.gov.pagopa.idpay.transactions.connector.rest.MerchantRestClient;
-import it.gov.pagopa.idpay.transactions.connector.rest.UserRestClient;
-import it.gov.pagopa.idpay.transactions.connector.rest.dto.MerchantDetailDTO;
-import it.gov.pagopa.idpay.transactions.connector.rest.erogazioni.ErogazioniRestClient;
-import it.gov.pagopa.idpay.transactions.connector.rest.invitalia.dto.ErogazioneOutcomeDTO;
-import it.gov.pagopa.idpay.transactions.connector.rest.invitalia.dto.InvitaliaOutcomeResponseDTO;
-import it.gov.pagopa.idpay.transactions.connector.rest.selfcare.SelfcareInstitutionsRestClient;
-import it.gov.pagopa.idpay.transactions.connector.rest.selfcare.dto.InstitutionDTO;
-import it.gov.pagopa.idpay.transactions.connector.rest.selfcare.dto.InstitutionList;
-import it.gov.pagopa.idpay.transactions.dto.ChecksErrorDTO;
-import it.gov.pagopa.idpay.transactions.dto.DeliveryRequest;
-import it.gov.pagopa.idpay.transactions.dto.TransactionsRequest;
-import it.gov.pagopa.idpay.transactions.dto.batch.BatchCountersDTO;
-import it.gov.pagopa.idpay.transactions.dto.mapper.ChecksErrorMapper;
-import it.gov.pagopa.idpay.transactions.enums.RewardBatchAssignee;
 import it.gov.pagopa.idpay.transactions.enums.RewardBatchStatus;
-import it.gov.pagopa.idpay.transactions.enums.RewardBatchTrxStatus;
-import it.gov.pagopa.idpay.transactions.model.ChecksError;
-import it.gov.pagopa.idpay.transactions.model.Reward;
 import it.gov.pagopa.idpay.transactions.model.RewardBatch;
-import it.gov.pagopa.idpay.transactions.model.RewardTransaction;
 import it.gov.pagopa.idpay.transactions.repository.RewardBatchRepository;
-import it.gov.pagopa.idpay.transactions.repository.RewardTransactionRepository;
-import it.gov.pagopa.idpay.transactions.storage.ApprovedRewardBatchBlobService;
-import it.gov.pagopa.idpay.transactions.utils.AuditUtilities;
-import it.gov.pagopa.idpay.transactions.utils.ExceptionConstants;
+import it.gov.pagopa.idpay.transactions.usecase.rewardbatch.RewardBatchSharedUtils;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.NullSource;
-import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.dao.DuplicateKeyException;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.mongodb.core.ReactiveMongoTemplate;
-import org.springframework.data.mongodb.core.query.Query;
-import org.springframework.http.HttpStatus;
-import org.springframework.test.util.ReflectionTestUtils;
-import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import reactor.test.StepVerifier;
 
-import java.lang.reflect.Method;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.YearMonth;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
 import static it.gov.pagopa.idpay.transactions.enums.PosType.PHYSICAL;
-import static it.gov.pagopa.idpay.transactions.utils.ExceptionConstants.ExceptionMessage.ERROR_MESSAGE_INVALID_CHECKS_ERROR;
-import static it.gov.pagopa.idpay.transactions.utils.ExceptionConstants.ExceptionMessage.MERCHANT_OR_OPERATOR_HEADER_MANDATORY;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.*;
@@ -70,44 +22,17 @@ import static org.mockito.Mockito.*;
 class RewardBatchServiceImplTest {
 
     @Mock private RewardBatchRepository rewardBatchRepository;
-    @Mock private RewardTransactionRepository rewardTransactionRepository;
-    @Mock private UserRestClient userRestClient;
-    @Mock private ApprovedRewardBatchBlobService approvedRewardBatchBlobService;
-    @Mock private ReactiveMongoTemplate reactiveMongoTemplate;
-    @Mock private ChecksErrorMapper checksErrorMapper;
-    @Mock private AuditUtilities auditUtilities;
-
-    @Mock private MerchantRestClient  merchantRestClient;
-    @Mock private SelfcareInstitutionsRestClient  selfcareInstitutionsRestClient;
-    @Mock private ErogazioniRestClient  erogazioniRestClient;
 
     private RewardBatchServiceImpl service;
-    private RewardBatchServiceImpl serviceSpy;
 
-    private static final String INITIATIVE_ID = "INITIATIVE_ID";
     private static final String MERCHANT_ID = "MERCHANT_ID";
     private static final String BUSINESS_NAME = "Business";
     private static final String BATCH_ID = "BATCH_ID";
     private static final String BATCH_ID_2 = "BATCH_ID_2";
-    private static final String OP1 = "operator1";
-    private static final String OP2 = "operator2";
-    private static final String OP3 = "operator3";
 
     @BeforeEach
     void setup() {
-        service = new RewardBatchServiceImpl(
-                rewardBatchRepository,
-                rewardTransactionRepository,
-                userRestClient,
-                approvedRewardBatchBlobService,
-                reactiveMongoTemplate,
-                checksErrorMapper,
-                auditUtilities,
-                merchantRestClient,
-                selfcareInstitutionsRestClient,
-                erogazioniRestClient
-        );
-        serviceSpy = spy(service);
+        service = new RewardBatchServiceImpl(rewardBatchRepository);
     }
 
     @Test
@@ -186,1755 +111,8 @@ class RewardBatchServiceImplTest {
     }
 
     @Test
-    void isOperator_privateRoleChecks() throws Exception {
-        Method m = RewardBatchServiceImpl.class.getDeclaredMethod("isOperator", String.class);
-        m.setAccessible(true);
-
-        assertFalse((boolean) m.invoke(service, (String) null));
-        assertFalse((boolean) m.invoke(service, "guest"));
-        assertTrue((boolean) m.invoke(service, "operator1"));
-        assertTrue((boolean) m.invoke(service, "OPERATOR2"));
-    }
-
-    @Test
-    void getRewardBatches_operatorVsMerchant() {
-        PageRequest pageable = PageRequest.of(0, 2);
-
-        RewardBatch b1 = RewardBatch.builder().id("B1").merchantId("M1").build();
-        RewardBatch b2 = RewardBatch.builder().id("B2").merchantId("M2").build();
-
-        when(rewardBatchRepository.findRewardBatchesCombined(null, null, null, null, true, pageable))
-                .thenReturn(Flux.just(b1, b2));
-        when(rewardBatchRepository.getCountCombined(null, null, null, null, true))
-                .thenReturn(Mono.just(10L));
-
-        StepVerifier.create(service.getRewardBatches(null, "operator1", null, null, null, pageable))
-                .assertNext(p -> {
-                    assertEquals(2, p.getContent().size());
-                    assertEquals(10L, p.getTotalElements());
-                })
-                .verifyComplete();
-
-        when(rewardBatchRepository.findRewardBatchesCombined("M1", null, null, null, false, pageable))
-                .thenReturn(Flux.just(b1));
-        when(rewardBatchRepository.getCountCombined("M1", null, null, null, false))
-                .thenReturn(Mono.just(1L));
-
-        StepVerifier.create(service.getRewardBatches("M1", "guest", null, null, null, pageable))
-                .assertNext(p -> {
-                    assertEquals(1, p.getContent().size());
-                    assertEquals(1L, p.getTotalElements());
-                })
-                .verifyComplete();
-    }
-
-
-    @Test
-    void sendRewardBatch_batchNotFound() {
-        when(rewardBatchRepository.findById(BATCH_ID)).thenReturn(Mono.empty());
-
-        StepVerifier.create(service.sendRewardBatch(MERCHANT_ID, BATCH_ID))
-                .expectError(RewardBatchException.class)
-                .verify();
-    }
-
-    @Test
-    void sendRewardBatch_merchantMismatch() {
-        RewardBatch b = RewardBatch.builder().id(BATCH_ID).merchantId("OTHER").status(RewardBatchStatus.CREATED).month("2025-01").posType(PHYSICAL).build();
-        when(rewardBatchRepository.findById(BATCH_ID)).thenReturn(Mono.just(b));
-
-        StepVerifier.create(service.sendRewardBatch(MERCHANT_ID, BATCH_ID))
-                .expectError(RewardBatchException.class)
-                .verify();
-    }
-
-    @Test
-    void sendRewardBatch_invalidStatus() {
-        RewardBatch b = RewardBatch.builder().id(BATCH_ID).merchantId(MERCHANT_ID).status(RewardBatchStatus.SENT).month("2025-01").posType(PHYSICAL).build();
-        when(rewardBatchRepository.findById(BATCH_ID)).thenReturn(Mono.just(b));
-
-        StepVerifier.create(service.sendRewardBatch(MERCHANT_ID, BATCH_ID))
-                .expectError(RewardBatchException.class)
-                .verify();
-    }
-
-    @Test
-    void sendRewardBatch_monthTooEarly() {
-        YearMonth now = YearMonth.now();
-        RewardBatch b = RewardBatch.builder().id(BATCH_ID).merchantId(MERCHANT_ID).status(RewardBatchStatus.CREATED).month(now.toString()).posType(PHYSICAL).build();
-        when(rewardBatchRepository.findById(BATCH_ID)).thenReturn(Mono.just(b));
-
-        StepVerifier.create(service.sendRewardBatch(MERCHANT_ID, BATCH_ID))
-                .expectErrorSatisfies(ex -> {
-                    assertInstanceOf(RewardBatchException.class, ex);
-                    assertTrue(ex.getMessage().contains("REWARD_BATCH_MONTH_TOO_EARLY"));
-                })
-                .verify();
-    }
-
-    @Test
-    void sendRewardBatch_previousNotSent() {
-        YearMonth batchMonth = YearMonth.now().minusMonths(1);
-
-        RewardBatch current = RewardBatch.builder()
-                .id(BATCH_ID)
-                .merchantId(MERCHANT_ID)
-                .status(RewardBatchStatus.CREATED)
-                .month(batchMonth.toString())
-                .posType(PHYSICAL)
-                .build();
-
-        RewardBatch previousCreated = RewardBatch.builder()
-                .id("PREV")
-                .merchantId(MERCHANT_ID)
-                .status(RewardBatchStatus.CREATED)
-                .month(batchMonth.minusMonths(1).toString())
-                .posType(PHYSICAL)
-                .build();
-
-        when(rewardBatchRepository.findById(BATCH_ID)).thenReturn(Mono.just(current));
-        when(rewardBatchRepository.findByMerchantIdAndPosType(MERCHANT_ID, PHYSICAL))
-                .thenReturn(Flux.just(previousCreated));
-
-        StepVerifier.create(service.sendRewardBatch(MERCHANT_ID, BATCH_ID))
-                .expectError(RewardBatchException.class)
-                .verify();
-    }
-
-    @Test
-    void sendRewardBatch_success_allPreviousSent() {
-        YearMonth batchMonth = YearMonth.now().minusMonths(1);
-
-        RewardBatch current = RewardBatch.builder()
-                .id(BATCH_ID)
-                .merchantId(MERCHANT_ID)
-                .status(RewardBatchStatus.CREATED)
-                .month(batchMonth.toString())
-                .posType(PHYSICAL)
-                .build();
-
-        RewardBatch previousSent = RewardBatch.builder()
-                .id("PREV")
-                .merchantId(MERCHANT_ID)
-                .status(RewardBatchStatus.SENT)
-                .month(batchMonth.minusMonths(1).toString())
-                .posType(PHYSICAL)
-                .build();
-
-        when(rewardBatchRepository.findById(BATCH_ID)).thenReturn(Mono.just(current));
-        when(rewardBatchRepository.findByMerchantIdAndPosType(MERCHANT_ID, PHYSICAL))
-                .thenReturn(Flux.just(previousSent));
-        when(rewardBatchRepository.save(any())).thenAnswer(inv -> Mono.just(inv.getArgument(0)));
-
-        StepVerifier.create(service.sendRewardBatch(MERCHANT_ID, BATCH_ID))
-                .verifyComplete();
-
-        verify(rewardBatchRepository).save(argThat(b -> b.getStatus() == RewardBatchStatus.SENT && b.getMerchantSendDate() != null));
-    }
-
-
-    @Test
-    void validChecksError_null_ok() {
-        assertDoesNotThrow(() -> serviceSpy.validChecksError(null));
-    }
-
-    @Test
-    void validChecksError_allFalse_throws() {
-        ChecksErrorDTO dto = new ChecksErrorDTO();
-        dto.setCfError(false);
-        dto.setProductEligibilityError(false);
-        dto.setDisposalRaeeError(false);
-        dto.setPriceError(false);
-        dto.setBonusError(false);
-        dto.setSellerReferenceError(false);
-        dto.setAccountingDocumentError(false);
-
-        InvalidChecksErrorException ex = assertThrows(InvalidChecksErrorException.class, () -> serviceSpy.validChecksError(dto));
-        assertEquals(ERROR_MESSAGE_INVALID_CHECKS_ERROR, ex.getMessage());
-    }
-
-    @Test
-    void validChecksError_anyTrue_ok() {
-        ChecksErrorDTO dto = new ChecksErrorDTO();
-        dto.setCfError(true);
-        assertDoesNotThrow(() -> serviceSpy.validChecksError(dto));
-    }
-
-    @Test
-    void validChecksError_nullDto_doesNothing() {
-        assertDoesNotThrow(() -> service.validChecksError(null));
-    }
-
-    @Test
-    void validChecksError_allFalse_throwsException() {
-        ChecksErrorDTO dto = new ChecksErrorDTO();
-
-        InvalidChecksErrorException ex =
-                assertThrows(InvalidChecksErrorException.class,
-                        () -> service.validChecksError(dto));
-
-        assertEquals(ERROR_MESSAGE_INVALID_CHECKS_ERROR, ex.getMessage());
-    }
-
-    @Test
-    void validChecksError_productEligibilityError_true_doesNotThrow() {
-        ChecksErrorDTO dto = new ChecksErrorDTO();
-        dto.setProductEligibilityError(true);
-
-        assertDoesNotThrow(() -> service.validChecksError(dto));
-    }
-
-    @Test
-    void validChecksError_disposalRaeeError_true_doesNotThrow() {
-        ChecksErrorDTO dto = new ChecksErrorDTO();
-        dto.setDisposalRaeeError(true);
-
-        assertDoesNotThrow(() -> service.validChecksError(dto));
-    }
-
-    @Test
-    void validChecksError_priceError_true_doesNotThrow() {
-        ChecksErrorDTO dto = new ChecksErrorDTO();
-        dto.setPriceError(true);
-
-        assertDoesNotThrow(() -> service.validChecksError(dto));
-    }
-
-    @Test
-    void validChecksError_bonusError_true_doesNotThrow() {
-        ChecksErrorDTO dto = new ChecksErrorDTO();
-        dto.setBonusError(true);
-
-        assertDoesNotThrow(() -> service.validChecksError(dto));
-    }
-
-    @Test
-    void validChecksError_sellerReferenceError_true_doesNotThrow() {
-        ChecksErrorDTO dto = new ChecksErrorDTO();
-        dto.setSellerReferenceError(true);
-
-        assertDoesNotThrow(() -> service.validChecksError(dto));
-    }
-
-    @Test
-    void validChecksError_accountingDocumentError_true_doesNotThrow() {
-        ChecksErrorDTO dto = new ChecksErrorDTO();
-        dto.setAccountingDocumentError(true);
-
-        assertDoesNotThrow(() -> service.validChecksError(dto));
-    }
-
-    @Test
-    void validChecksError_genericError_true_doesNotThrow() {
-        ChecksErrorDTO dto = new ChecksErrorDTO();
-        dto.setGenericError(true);
-
-        assertDoesNotThrow(() -> service.validChecksError(dto));
-    }
-
-    @Test
-    void suspendTransactions_batchNotFoundOrInvalidState() {
-        TransactionsRequest req = TransactionsRequest.builder().transactionIds(List.of("t1")).checksError(new ChecksErrorDTO(){{
-            setCfError(true);
-        }}).build();
-
-        when(checksErrorMapper.toModel(any())).thenReturn(new ChecksError());
-        when(rewardBatchRepository.findByIdAndStatus(BATCH_ID, RewardBatchStatus.EVALUATING))
-                .thenReturn(Mono.empty());
-
-        StepVerifier.create(service.suspendTransactions(BATCH_ID, INITIATIVE_ID, req))
-                .expectError(ClientExceptionWithBody.class)
-                .verify();
-
-        verify(rewardTransactionRepository, never()).updateStatusAndReturnOld(any(), any(), any(), any(), any(), any());
-    }
-
-    @Test
-    void suspendTransactions_coversAllBranches() {
-        String batchMonth = "2025-12";
-        RewardBatch batch = RewardBatch.builder().id(BATCH_ID).status(RewardBatchStatus.EVALUATING).month(batchMonth).build();
-
-        ChecksErrorDTO checks = new ChecksErrorDTO();
-        checks.setCfError(true);
-
-        TransactionsRequest req = TransactionsRequest.builder()
-                .transactionIds(List.of("SUSP_PREV", "APP", "TO_CHECK", "CONS", "REJ", "NULL_ACC"))
-                .reason("REASON")
-                .checksError(checks)
-                .build();
-
-        ChecksError model = new ChecksError();
-        when(checksErrorMapper.toModel(checks)).thenReturn(model);
-
-        when(rewardBatchRepository.findByIdAndStatus(BATCH_ID, RewardBatchStatus.EVALUATING))
-                .thenReturn(Mono.just(batch));
-
-        RewardTransaction trxSuspPrev = RewardTransaction.builder()
-                .id("SUSP_PREV")
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.SUSPENDED)
-                .rewardBatchLastMonthElaborated("2025-11")
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(100L).build()))
-                .build();
-
-        RewardTransaction trxApproved = RewardTransaction.builder()
-                .id("APP")
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.APPROVED)
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(200L).build()))
-                .build();
-
-        RewardTransaction trxToCheck = RewardTransaction.builder()
-                .id("TO_CHECK")
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.TO_CHECK)
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(300L).build()))
-                .build();
-
-        RewardTransaction trxConsultable = RewardTransaction.builder()
-                .id("CONS")
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.CONSULTABLE)
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(400L).build()))
-                .build();
-
-        RewardTransaction trxRejected = RewardTransaction.builder()
-                .id("REJ")
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.REJECTED)
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(500L).build()))
-                .build();
-
-        RewardTransaction trxNullAccrued = RewardTransaction.builder()
-                .id("NULL_ACC")
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.APPROVED)
-                .rewards(Map.of("OTHER", Reward.builder().accruedRewardCents(999L).build()))
-                .build();
-
-        when(rewardTransactionRepository.updateStatusAndReturnOld(eq(BATCH_ID), eq("SUSP_PREV"), eq(RewardBatchTrxStatus.SUSPENDED), any(), eq(batchMonth), eq(model)))
-                .thenReturn(Mono.just(trxSuspPrev));
-        when(rewardTransactionRepository.updateStatusAndReturnOld(eq(BATCH_ID), eq("APP"), eq(RewardBatchTrxStatus.SUSPENDED), any(), eq(batchMonth),eq( model)))
-                .thenReturn(Mono.just(trxApproved));
-        when(rewardTransactionRepository.updateStatusAndReturnOld(eq(BATCH_ID), eq("TO_CHECK"), eq(RewardBatchTrxStatus.SUSPENDED), any(), eq(batchMonth), eq(model)))
-                .thenReturn(Mono.just(trxToCheck));
-        when(rewardTransactionRepository.updateStatusAndReturnOld(eq(BATCH_ID), eq("CONS"), eq(RewardBatchTrxStatus.SUSPENDED), any(), eq(batchMonth), eq(model)))
-                .thenReturn(Mono.just(trxConsultable));
-        when(rewardTransactionRepository.updateStatusAndReturnOld(eq(BATCH_ID), eq("REJ"), eq(RewardBatchTrxStatus.SUSPENDED), any(), eq(batchMonth), eq(model)))
-                .thenReturn(Mono.just(trxRejected));
-        when(rewardTransactionRepository.updateStatusAndReturnOld(eq(BATCH_ID), eq("NULL_ACC"),eq( RewardBatchTrxStatus.SUSPENDED), any(), eq(batchMonth), eq(model)))
-                .thenReturn(Mono.just(trxNullAccrued));
-
-        RewardBatch updated = RewardBatch.builder().id(BATCH_ID).build();
-        when(rewardBatchRepository.updateTotals(eq(BATCH_ID), any(BatchCountersDTO.class)))
-                .thenReturn(Mono.just(updated));
-
-        StepVerifier.create(service.suspendTransactions(BATCH_ID, INITIATIVE_ID, req))
-                .expectNext(updated)
-                .verifyComplete();
-
-        verify(auditUtilities).logTransactionsStatusChanged(eq(RewardBatchTrxStatus.SUSPENDED.name()), eq(INITIATIVE_ID), anyString(), eq(checks));
-        verify(rewardBatchRepository).updateTotals(eq(BATCH_ID), any(BatchCountersDTO.class));
-    }
-
-    @Test
-    void suspendTransactions_alreadySuspended_sameMonth_skipsElaboratedIncrement() {
-        String batchMonth = "2025-12";
-        RewardBatch batch = RewardBatch.builder().id(BATCH_ID).status(RewardBatchStatus.EVALUATING).month(batchMonth).build();
-
-        ChecksErrorDTO checks = new ChecksErrorDTO();
-        checks.setCfError(true);
-
-        TransactionsRequest req = TransactionsRequest.builder()
-                .transactionIds(List.of("SUSP_SAME"))
-                .reason("R")
-                .checksError(checks)
-                .build();
-
-        ChecksError model = new ChecksError();
-        when(checksErrorMapper.toModel(checks)).thenReturn(model);
-        when(rewardBatchRepository.findByIdAndStatus(BATCH_ID, RewardBatchStatus.EVALUATING))
-                .thenReturn(Mono.just(batch));
-
-        RewardTransaction trxSuspSame = RewardTransaction.builder()
-                .id("SUSP_SAME")
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.SUSPENDED)
-                .rewardBatchLastMonthElaborated("2025-12")
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(100L).build()))
-                .build();
-
-        when(rewardTransactionRepository.updateStatusAndReturnOld(eq(BATCH_ID), eq("SUSP_SAME"), eq(RewardBatchTrxStatus.SUSPENDED), any(), eq(batchMonth), eq(model)))
-                .thenReturn(Mono.just(trxSuspSame));
-
-        when(rewardBatchRepository.updateTotals(eq(BATCH_ID), any(BatchCountersDTO.class)))
-                .thenReturn(Mono.just(batch));
-
-        StepVerifier.create(service.suspendTransactions(BATCH_ID, INITIATIVE_ID, req))
-                .expectNext(batch)
-                .verifyComplete();
-    }
-
-    @Test
-    void rejectTransactions_allBranches() {
-        String batchMonth = "2025-12";
-        RewardBatch batch = RewardBatch.builder().id(BATCH_ID).status(RewardBatchStatus.EVALUATING).month(batchMonth).build();
-
-        TransactionsRequest req = TransactionsRequest.builder()
-                .transactionIds(List.of("ALREADY_REJ", "APP", "TO_CHECK", "CONS", "SUSP_PREV"))
-                .reason("WHY")
-                .build();
-
-        when(rewardBatchRepository.findByIdAndStatus(BATCH_ID, RewardBatchStatus.EVALUATING))
-                .thenReturn(Mono.just(batch));
-
-        RewardTransaction alreadyRejected = RewardTransaction.builder()
-                .id("ALREADY_REJ")
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.REJECTED)
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(10L).build()))
-                .build();
-
-        RewardTransaction approved = RewardTransaction.builder()
-                .id("APP")
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.APPROVED)
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(20L).build()))
-                .build();
-
-        RewardTransaction toCheck = RewardTransaction.builder()
-                .id("TO_CHECK")
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.TO_CHECK)
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(30L).build()))
-                .build();
-
-        RewardTransaction consultable = RewardTransaction.builder()
-                .id("CONS")
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.CONSULTABLE)
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(40L).build()))
-                .build();
-
-        RewardTransaction suspendedPrev = RewardTransaction.builder()
-                .id("SUSP_PREV")
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.SUSPENDED)
-                .rewardBatchLastMonthElaborated("2025-11")
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(50L).build()))
-                .build();
-
-        when(rewardTransactionRepository.updateStatusAndReturnOld(eq(BATCH_ID), eq("ALREADY_REJ"), eq(RewardBatchTrxStatus.REJECTED), any(),eq(batchMonth), eq(null)))
-                .thenReturn(Mono.just(alreadyRejected));
-        when(rewardTransactionRepository.updateStatusAndReturnOld(eq(BATCH_ID), eq("APP"), eq(RewardBatchTrxStatus.REJECTED), any(),eq(batchMonth), eq(null)))
-                .thenReturn(Mono.just(approved));
-        when(rewardTransactionRepository.updateStatusAndReturnOld(eq(BATCH_ID), eq("TO_CHECK"), eq(RewardBatchTrxStatus.REJECTED), any(), eq(batchMonth), eq(null)))
-                .thenReturn(Mono.just(toCheck));
-        when(rewardTransactionRepository.updateStatusAndReturnOld(eq(BATCH_ID), eq("CONS"), eq(RewardBatchTrxStatus.REJECTED), any(), eq(batchMonth), eq(null)))
-                .thenReturn(Mono.just(consultable));
-        when(rewardTransactionRepository.updateStatusAndReturnOld(eq(BATCH_ID), eq("SUSP_PREV"), eq(RewardBatchTrxStatus.REJECTED), any(), eq(batchMonth), eq(null)))
-                .thenReturn(Mono.just(suspendedPrev));
-
-        RewardBatch updated = RewardBatch.builder().id(BATCH_ID).build();
-        when(rewardBatchRepository.updateTotals(eq(BATCH_ID), any(BatchCountersDTO.class)))
-                .thenReturn(Mono.just(updated));
-
-        StepVerifier.create(service.rejectTransactions(BATCH_ID, INITIATIVE_ID, req))
-                .expectNext(updated)
-                .verifyComplete();
-    }
-
-    @Test
-    void approvedTransactions_allBranches() {
-        String batchMonth = "2025-12";
-        RewardBatch batch = RewardBatch.builder().id(BATCH_ID).status(RewardBatchStatus.EVALUATING).month(batchMonth).build();
-
-        TransactionsRequest req = TransactionsRequest.builder()
-                .transactionIds(List.of("ALREADY_APP", "TO_CHECK", "CONS", "SUSP_PREV", "REJ"))
-                .build();
-
-        when(rewardBatchRepository.findByIdAndStatus(BATCH_ID, RewardBatchStatus.EVALUATING))
-                .thenReturn(Mono.just(batch));
-
-        RewardTransaction alreadyApproved = RewardTransaction.builder()
-                .id("ALREADY_APP")
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.APPROVED)
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(10L).build()))
-                .build();
-
-        RewardTransaction toCheck = RewardTransaction.builder()
-                .id("TO_CHECK")
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.TO_CHECK)
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(20L).build()))
-                .build();
-
-        RewardTransaction consultable = RewardTransaction.builder()
-                .id("CONS")
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.CONSULTABLE)
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(30L).build()))
-                .build();
-
-        RewardTransaction suspendedPrev = RewardTransaction.builder()
-                .id("SUSP_PREV")
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.SUSPENDED)
-                .rewardBatchLastMonthElaborated("2025-11")
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(40L).build()))
-                .build();
-
-        RewardTransaction rejected = RewardTransaction.builder()
-                .id("REJ")
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.REJECTED)
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(50L).build()))
-                .build();
-
-        when(rewardTransactionRepository.updateStatusAndReturnOld(BATCH_ID, "ALREADY_APP", RewardBatchTrxStatus.APPROVED, null, batchMonth, null))
-                .thenReturn(Mono.just(alreadyApproved));
-        when(rewardTransactionRepository.updateStatusAndReturnOld(BATCH_ID, "TO_CHECK", RewardBatchTrxStatus.APPROVED, null, batchMonth, null))
-                .thenReturn(Mono.just(toCheck));
-        when(rewardTransactionRepository.updateStatusAndReturnOld(BATCH_ID, "CONS", RewardBatchTrxStatus.APPROVED, null, batchMonth, null))
-                .thenReturn(Mono.just(consultable));
-        when(rewardTransactionRepository.updateStatusAndReturnOld(BATCH_ID, "SUSP_PREV", RewardBatchTrxStatus.APPROVED, null, batchMonth, null))
-                .thenReturn(Mono.just(suspendedPrev));
-        when(rewardTransactionRepository.updateStatusAndReturnOld(BATCH_ID, "REJ", RewardBatchTrxStatus.APPROVED, null, batchMonth, null))
-                .thenReturn(Mono.just(rejected));
-
-        RewardBatch updated = RewardBatch.builder().id(BATCH_ID).build();
-        when(rewardBatchRepository.updateTotals(eq(BATCH_ID), any(BatchCountersDTO.class)))
-                .thenReturn(Mono.just(updated));
-
-        StepVerifier.create(service.approvedTransactions(BATCH_ID, req, INITIATIVE_ID))
-                .expectNext(updated)
-                .verifyComplete();
-    }
-
-    @Test
-    void approvedTransactions_batchNotFoundOrInvalidState() {
-        when(rewardBatchRepository.findByIdAndStatus(BATCH_ID, RewardBatchStatus.EVALUATING))
-                .thenReturn(Mono.empty());
-
-        TransactionsRequest req = TransactionsRequest.builder().transactionIds(List.of("t1")).build();
-
-        StepVerifier.create(service.approvedTransactions(BATCH_ID, req, INITIATIVE_ID))
-                .expectError(ClientExceptionWithBody.class)
-                .verify();
-    }
-
-    @Test
-    void evaluatingRewardBatches_nullList_processesAllSent() {
-        RewardBatch sent = RewardBatch.builder().id("S1")
-                .status(RewardBatchStatus.SENT)
-                .initialAmountCents(100L)
-                .suspendedAmountCents(0L).build();
-
-        when(rewardBatchRepository.findByStatus(RewardBatchStatus.SENT)).thenReturn(Flux.just(sent));
-        when(rewardTransactionRepository.rewardTransactionsByBatchId("S1")).thenReturn(Mono.empty());
-        when(rewardTransactionRepository.sumSuspendedAccruedRewardCents("S1")).thenReturn(Mono.just(20L));
-        when(rewardBatchRepository.updateStatusAndApprovedAmountCents("S1", RewardBatchStatus.EVALUATING, 100L))
-                .thenReturn(Mono.just(sent));
-
-        StepVerifier.create(service.evaluatingRewardBatches(null))
-                .expectNext(1L)
-                .verifyComplete();
-    }
-
-    @Test
-    void evaluatingRewardBatches_withList_handlesMissingIdsAsEmpty() {
-        when(rewardBatchRepository.findByIdAndStatus("S1", RewardBatchStatus.SENT)).thenReturn(Mono.empty());
-
-        StepVerifier.create(service.evaluatingRewardBatches(List.of("S1")))
-                .expectNext(0L)
-                .verifyComplete();
-
-        verify(rewardBatchRepository, never()).updateStatusAndApprovedAmountCents(any(), any(), anyLong());
-    }
-
-
-
-    @Test
-    void downloadApprovedRewardBatchFile_invalidRequest_missingHeaders() {
-        StepVerifier.create(service.downloadApprovedRewardBatchFile(null, null, INITIATIVE_ID, BATCH_ID))
-                .expectErrorSatisfies(ex -> {
-                    assertInstanceOf(RewardBatchInvalidRequestException.class, ex);
-                    assertEquals(MERCHANT_OR_OPERATOR_HEADER_MANDATORY, ex.getMessage());
-                })
-                .verify();
-    }
-
-    @Test
-    void downloadApprovedRewardBatchFile_notFound_merchantPath() {
-        when(rewardBatchRepository.findByMerchantIdAndId(MERCHANT_ID, BATCH_ID)).thenReturn(Mono.empty());
-
-        StepVerifier.create(service.downloadApprovedRewardBatchFile(MERCHANT_ID, OP1, INITIATIVE_ID, BATCH_ID))
-                .expectError(RewardBatchNotFound.class)
-                .verify();
-    }
-
-    @Test
-    void downloadApprovedRewardBatchFile_roleNotAllowed_whenMerchantNull() {
-        RewardBatch batch = RewardBatch.builder().id(BATCH_ID).status(RewardBatchStatus.APPROVED).filename("file.csv").merchantId(MERCHANT_ID).build();
-        when(rewardBatchRepository.findById(BATCH_ID)).thenReturn(Mono.just(batch));
-
-        StepVerifier.create(service.downloadApprovedRewardBatchFile(null, "guest", INITIATIVE_ID, BATCH_ID))
-                .expectError(RoleNotAllowedException.class)
-                .verify();
-    }
-
-    @Test
-    void downloadApprovedRewardBatchFile_notApproved() {
-        RewardBatch batch = RewardBatch.builder().id(BATCH_ID).status(RewardBatchStatus.EVALUATING).filename("file.csv").merchantId(MERCHANT_ID).build();
-        when(rewardBatchRepository.findByMerchantIdAndId(MERCHANT_ID, BATCH_ID)).thenReturn(Mono.just(batch));
-
-        StepVerifier.create(service.downloadApprovedRewardBatchFile(MERCHANT_ID, OP1, INITIATIVE_ID, BATCH_ID))
-                .expectError(RewardBatchNotApprovedException.class)
-                .verify();
-    }
-
-    @Test
-    void downloadApprovedRewardBatchFile_allowed_whenRefundState() {
-
-        RewardBatch batch = RewardBatch.builder()
-                .id(BATCH_ID)
-                .status(RewardBatchStatus.PENDING_REFUND)
-                .filename("file.csv")
-                .merchantId(MERCHANT_ID)
-                .build();
-
-        when(rewardBatchRepository.findByMerchantIdAndId(MERCHANT_ID, BATCH_ID))
-                .thenReturn(Mono.just(batch));
-
-        when(approvedRewardBatchBlobService.getFileSignedUrl(anyString()))
-                .thenReturn("signed-refund");
-
-        StepVerifier.create(
-                        service.downloadApprovedRewardBatchFile(
-                                MERCHANT_ID,
-                                OP1,
-                                INITIATIVE_ID,
-                                BATCH_ID
-                        )
-                )
-                .assertNext(r -> assertEquals("signed-refund", r.getApprovedBatchUrl()))
-                .verifyComplete();
-    }
-
-    @ParameterizedTest
-    @NullSource
-    @ValueSource(strings = {"", "   "})
-    void downloadApprovedRewardBatchFile_missingFilename(String invalid) {
-        RewardBatch batch = RewardBatch.builder().id(BATCH_ID).status(RewardBatchStatus.APPROVED).filename(invalid).merchantId(MERCHANT_ID).build();
-        when(rewardBatchRepository.findByMerchantIdAndId(MERCHANT_ID, BATCH_ID)).thenReturn(Mono.just(batch));
-
-        StepVerifier.create(service.downloadApprovedRewardBatchFile(MERCHANT_ID, OP1, INITIATIVE_ID, BATCH_ID))
-                .expectError(RewardBatchMissingFilenameException.class)
-                .verify();
-    }
-
-    @Test
-    void downloadApprovedRewardBatchFile_success_merchant() {
-        RewardBatch batch = RewardBatch.builder()
-                .id(BATCH_ID).status(RewardBatchStatus.APPROVED).filename("file.csv").merchantId(MERCHANT_ID).build();
-
-        when(rewardBatchRepository.findByMerchantIdAndId(MERCHANT_ID, BATCH_ID)).thenReturn(Mono.just(batch));
-        when(approvedRewardBatchBlobService.getFileSignedUrl(anyString())).thenReturn("signed");
-
-        StepVerifier.create(service.downloadApprovedRewardBatchFile(MERCHANT_ID, OP1, INITIATIVE_ID, BATCH_ID))
-                .assertNext(r -> assertEquals("signed", r.getApprovedBatchUrl()))
-                .verifyComplete();
-    }
-
-    @Test
-    void downloadApprovedRewardBatchFile_success_operatorOnlyMerchantNull() {
-        RewardBatch batch = RewardBatch.builder()
-                .id(BATCH_ID).status(RewardBatchStatus.APPROVED).filename("file.csv").merchantId(MERCHANT_ID).build();
-
-        when(rewardBatchRepository.findById(BATCH_ID)).thenReturn(Mono.just(batch));
-        when(approvedRewardBatchBlobService.getFileSignedUrl(anyString())).thenReturn("signed2");
-
-        StepVerifier.create(service.downloadApprovedRewardBatchFile(null, OP2, INITIATIVE_ID, BATCH_ID))
-                .assertNext(r -> assertEquals("signed2", r.getApprovedBatchUrl()))
-                .verifyComplete();
-    }
-
-
-    @Test
-    void rewardBatchConfirmation_notFound() {
-        when(rewardBatchRepository.findRewardBatchById(BATCH_ID)).thenReturn(Mono.empty());
-
-        StepVerifier.create(service.rewardBatchConfirmation(INITIATIVE_ID, BATCH_ID))
-                .expectError(ClientExceptionWithBody.class)
-                .verify();
-    }
-
-    @Test
-    void rewardBatchConfirmation_invalidState() {
-        RewardBatch rb = RewardBatch.builder().id(BATCH_ID).status(RewardBatchStatus.EVALUATING).assigneeLevel(RewardBatchAssignee.L2).build();
-        when(rewardBatchRepository.findRewardBatchById(BATCH_ID)).thenReturn(Mono.just(rb));
-
-        StepVerifier.create(service.rewardBatchConfirmation(INITIATIVE_ID, BATCH_ID))
-                .expectError(ClientExceptionWithBody.class)
-                .verify();
-    }
-
-    @Test
-    void rewardBatchConfirmation_previousNotApproved_blocks() {
-        RewardBatch rb = RewardBatch.builder().id(BATCH_ID).status(RewardBatchStatus.EVALUATING).assigneeLevel(RewardBatchAssignee.L3)
-                .merchantId(MERCHANT_ID).posType(PHYSICAL).month("2025-12").build();
-
-        RewardBatch prev = RewardBatch.builder().id("P1").status(RewardBatchStatus.SENT).build();
-
-        when(rewardBatchRepository.findRewardBatchById(BATCH_ID)).thenReturn(Mono.just(rb));
-        when(rewardBatchRepository.findRewardBatchByMonthBefore(MERCHANT_ID, PHYSICAL, "2025-12"))
-                .thenReturn(Flux.just(prev));
-
-        StepVerifier.create(service.rewardBatchConfirmation(INITIATIVE_ID, BATCH_ID))
-                .expectError(ClientExceptionWithBody.class)
-                .verify();
-    }
-
-    @Test
-    void rewardBatchConfirmation_success() {
-        RewardBatch rb = RewardBatch.builder().id(BATCH_ID).status(RewardBatchStatus.EVALUATING).assigneeLevel(RewardBatchAssignee.L3)
-                .merchantId(MERCHANT_ID).posType(PHYSICAL).month("2025-12").build();
-
-        RewardBatch prevApproved = RewardBatch.builder().id("P1").status(RewardBatchStatus.APPROVED).build();
-
-        when(rewardBatchRepository.findRewardBatchById(BATCH_ID)).thenReturn(Mono.just(rb));
-        when(rewardBatchRepository.findRewardBatchByMonthBefore(MERCHANT_ID, PHYSICAL, "2025-12"))
-                .thenReturn(Flux.just(prevApproved));
-        when(rewardBatchRepository.save(any())).thenAnswer(inv -> Mono.just(inv.getArgument(0)));
-
-        StepVerifier.create(service.rewardBatchConfirmation(INITIATIVE_ID, BATCH_ID))
-                .assertNext(updated -> {
-                    assertEquals(RewardBatchStatus.APPROVING, updated.getStatus());
-                    assertNotNull(updated.getApprovalDate());
-                    assertNotNull(updated.getUpdateDate());
-                })
-                .verifyComplete();
-    }
-
-    @Test
-    void rewardBatchConfirmationBatch_withIds_processesEach() {
-        doReturn(Mono.just(new RewardBatch())).when(serviceSpy).processSingleBatchConfirmation(eq(BATCH_ID), anyString());
-        doReturn(Mono.just(new RewardBatch())).when(serviceSpy).processSingleBatchConfirmation(eq(BATCH_ID_2), anyString());
-
-        StepVerifier.create(serviceSpy.rewardBatchConfirmationBatch(INITIATIVE_ID, List.of(BATCH_ID, BATCH_ID_2)))
-                .verifyComplete();
-
-        verify(serviceSpy).processSingleBatchConfirmation(BATCH_ID, INITIATIVE_ID);
-        verify(serviceSpy).processSingleBatchConfirmation(BATCH_ID_2, INITIATIVE_ID);
-        verify(rewardBatchRepository, never()).findRewardBatchByStatus(any());
-    }
-
-    @Test
-    void rewardBatchConfirmationBatch_emptyList_fetchesApprovingAndProcesses() {
-        RewardBatch b1 = RewardBatch.builder().id(BATCH_ID).status(RewardBatchStatus.APPROVING).build();
-
-        when(rewardBatchRepository.findRewardBatchByStatus(RewardBatchStatus.APPROVING))
-                .thenReturn(Flux.just(b1));
-        doReturn(Mono.just(b1)).when(serviceSpy).processSingleBatchConfirmation(eq(BATCH_ID), anyString());
-
-        StepVerifier.create(serviceSpy.rewardBatchConfirmationBatch(INITIATIVE_ID, Collections.emptyList()))
-                .verifyComplete();
-
-        verify(serviceSpy).processSingleBatchConfirmation(BATCH_ID, INITIATIVE_ID);
-    }
-
-
-    @Test
-    void processBatchesOrchestrator_shouldContinueOnSingleBatchError() {
-        doReturn(Mono.error(new RuntimeException("Error Batch 1")))
-                .when(serviceSpy).processSingleBatchDelivery(eq(BATCH_ID), anyString());
-        doReturn(Mono.just(new RewardBatch()))
-                .when(serviceSpy).processSingleBatchDelivery(eq(BATCH_ID_2), anyString());
-
-        StepVerifier.create(serviceSpy.rewardBatchDeliveryBatch(INITIATIVE_ID, List.of(BATCH_ID, BATCH_ID_2)))
-                .verifyComplete();
-
-        verify(serviceSpy).processSingleBatchDelivery(BATCH_ID, INITIATIVE_ID);
-        verify(serviceSpy).processSingleBatchDelivery(BATCH_ID_2, INITIATIVE_ID);
-    }
-
-    @Test
-    void rewardBatchDeliveryBatch_Success() {
-        String initiativeId = "INIT_1";
-        String batchId = "BATCH_1";
-        String merchantId = "MERCHANT_1";
-        String fiscalCode = "FISCAL_123";
-
-        RewardBatch batch = new RewardBatch();
-        batch.setId(batchId);
-        batch.setMerchantId(merchantId);
-        batch.setStatus(RewardBatchStatus.APPROVED);
-        batch.setApprovedAmountCents(1000L);
-
-        MerchantDetailDTO merchantDetail = new MerchantDetailDTO();
-        merchantDetail.setFiscalCode(fiscalCode);
-        merchantDetail.setVatNumber("VAT_123");
-
-        InstitutionDTO inst = new InstitutionDTO();
-        inst.setZipCode("00100");
-        inst.setDigitalAddress("pec@test.it");
-        InstitutionList instList = new InstitutionList(List.of(inst));
-
-        when(rewardBatchRepository.findRewardBatchById(batchId)).thenReturn(Mono.just(batch));
-        when(merchantRestClient.getMerchantDetail(merchantId, initiativeId)).thenReturn(Mono.just(merchantDetail));
-        when(selfcareInstitutionsRestClient.getInstitutions(fiscalCode)).thenReturn(Mono.just(instList));
-        when(erogazioniRestClient.postErogazione(any(DeliveryRequest.class))).thenReturn(Mono.empty());
-
-        StepVerifier.create(serviceSpy.rewardBatchDeliveryBatch(initiativeId, List.of(batchId)))
-                .verifyComplete();
-
-        verify(erogazioniRestClient).postErogazione(argThat(req ->
-                req.getId().equals(batchId) && req.getAnagrafica().getCap().equals("00100")
-        ));
-    }
-
-    @Test
-    void rewardBatchDeliveryBatch_Fail_MultipleInstitutions() {
-        // Given
-        String initiativeId = "INIT_1";
-        String batchId = "BATCH_1";
-        String fiscalCode = "FISCAL_123";
-
-        RewardBatch batch = new RewardBatch();
-        batch.setId(batchId);
-        batch.setMerchantId("M1");
-        batch.setStatus(RewardBatchStatus.APPROVED);
-
-        MerchantDetailDTO merchantDetail = new MerchantDetailDTO();
-        merchantDetail.setFiscalCode(fiscalCode);
-        merchantDetail.setVatNumber("VAT123");
-        merchantDetail.setBusinessName("Business");
-        merchantDetail.setIban("IT00TEST");
-        merchantDetail.setIbanHolder("Holder");
-
-        InstitutionList instList = new InstitutionList(List.of(new InstitutionDTO(), new InstitutionDTO()));
-
-        when(rewardBatchRepository.findRewardBatchById(batchId)).thenReturn(Mono.just(batch));
-        when(merchantRestClient.getMerchantDetail(anyString(), anyString())).thenReturn(Mono.just(merchantDetail));
-        when(selfcareInstitutionsRestClient.getInstitutions(fiscalCode)).thenReturn(Mono.just(instList));
-
-        StepVerifier.create(service.rewardBatchDeliveryBatch(initiativeId, List.of(batchId)))
-                .verifyComplete();
-
-        verify(erogazioniRestClient, never()).postErogazione(any());
-    }
-
-    @Test
-    void handleSuspendedTransactions_nullOrZero_returnsOriginal() {
-        RewardBatch rbNull = RewardBatch.builder().id(BATCH_ID).numberOfTransactionsSuspended(null).build();
-        RewardBatch rbZero = RewardBatch.builder().id(BATCH_ID).numberOfTransactionsSuspended(0L).build();
-
-        Mono<RewardBatch> r1 = ReflectionTestUtils.invokeMethod(serviceSpy, "handleSuspendedTransactions", rbNull, INITIATIVE_ID);
-        Mono<RewardBatch> r2 = ReflectionTestUtils.invokeMethod(serviceSpy, "handleSuspendedTransactions", rbZero, INITIATIVE_ID);
-
-        assertNotNull(r1);
-        StepVerifier.create(r1).expectNext(rbNull).verifyComplete();
-        assertNotNull(r2);
-        StepVerifier.create(r2).expectNext(rbZero).verifyComplete();
-    }
-
-    @Test
-    void createRewardBatchAndSave_existingBatchFound() {
-        RewardBatch savedBatch = RewardBatch.builder()
-                .id(BATCH_ID)
-                .merchantId(MERCHANT_ID)
-                .businessName(BUSINESS_NAME)
-                .month("2025-12")
-                .name("dicembre 2025")
-                .posType(PHYSICAL)
-                .status(RewardBatchStatus.APPROVED)
-                .partial(false)
-                .build();
-
-        RewardBatch existing = RewardBatch.builder().id(BATCH_ID_2).merchantId(MERCHANT_ID).month("2026-01").name("gennaio 2026").posType(PHYSICAL).build();
-
-        when(rewardBatchRepository.findRewardBatchByFilter(null, MERCHANT_ID, PHYSICAL, "2026-01"))
-                .thenReturn(Mono.just(existing));
-
-        StepVerifier.create(service.createRewardBatchAndSave(savedBatch))
-                .expectNext(existing)
-                .verifyComplete();
-
-        verify(rewardBatchRepository, never()).save(argThat(b -> b.getId() == null));
-    }
-
-    @Test
-    void createRewardBatchAndSave_createsWhenMissing() {
-        RewardBatch savedBatch = RewardBatch.builder()
-                .id(BATCH_ID)
-                .merchantId(MERCHANT_ID)
-                .businessName(BUSINESS_NAME)
-                .month("2025-12")
-                .name("dicembre 2025")
-                .posType(PHYSICAL)
-                .partial(false)
-                .build();
-
-        when(rewardBatchRepository.findRewardBatchByFilter(null, MERCHANT_ID, PHYSICAL, "2026-01"))
-                .thenReturn(Mono.empty());
-
-        when(rewardBatchRepository.save(any()))
-                .thenAnswer(inv -> {
-                    RewardBatch b = inv.getArgument(0);
-                    b.setId(BATCH_ID_2);
-                    return Mono.just(b);
-                });
-
-        StepVerifier.create(service.createRewardBatchAndSave(savedBatch))
-                .assertNext(b -> {
-                    assertEquals(BATCH_ID_2, b.getId());
-                    assertEquals("2026-01", b.getMonth());
-                    assertEquals(RewardBatchStatus.CREATED, b.getStatus());
-                    assertEquals(RewardBatchAssignee.L1, b.getAssigneeLevel());
-                })
-                .verifyComplete();
-
-        verify(rewardBatchRepository).save(any());
-    }
-
-    @Test
     void addOneMonth_and_italian() {
-        assertEquals("2026-01", service.addOneMonth("2025-12"));
-        assertEquals("gennaio 2026", service.addOneMonthToItalian("dicembre 2025"));
-    }
-
-    @Test
-    void updateAndSaveRewardTransactionsToApprove_setsApprovedAndSaves() {
-        RewardTransaction t1 = new RewardTransaction();
-        t1.setId("T1");
-        t1.setRewardBatchTrxStatus(RewardBatchTrxStatus.TO_CHECK);
-
-        when(rewardTransactionRepository.findByFilter(eq(BATCH_ID), eq(INITIATIVE_ID), anyList()))
-                .thenReturn(Flux.just(t1));
-        when(rewardTransactionRepository.save(any())).thenAnswer(inv -> Mono.just(inv.getArgument(0)));
-
-        StepVerifier.create(service.updateAndSaveRewardTransactionsToApprove(BATCH_ID, INITIATIVE_ID))
-                .verifyComplete();
-
-        verify(rewardTransactionRepository).save(argThat(t -> t.getRewardBatchTrxStatus() == RewardBatchTrxStatus.APPROVED));
-    }
-
-    @Test
-    void updateAndSaveRewardTransactionsSuspended_empty_returnsZero() {
-        when(rewardTransactionRepository.findByFilter(eq(BATCH_ID), eq(INITIATIVE_ID), anyList()))
-                .thenReturn(Flux.empty());
-
-        StepVerifier.create(service.updateAndSaveRewardTransactionsSuspended(BATCH_ID, INITIATIVE_ID, BATCH_ID_2, "2025-12"))
-                .expectNext(0L)
-                .verifyComplete();
-    }
-
-    @Test
-    void updateAndSaveRewardTransactionsSuspended_movesAndSums_andSetsLastMonthIfNull() {
-        RewardTransaction t1 = RewardTransaction.builder()
-                .id("T1")
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.SUSPENDED)
-                .rewardBatchId(BATCH_ID)
-                .rewardBatchLastMonthElaborated(null)
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(100L).build()))
-                .build();
-
-        RewardTransaction t2 = RewardTransaction.builder()
-                .id("T2")
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.SUSPENDED)
-                .rewardBatchId(BATCH_ID)
-                .rewardBatchLastMonthElaborated("2025-11")
-                .rewards(null)
-                .build();
-
-        when(rewardTransactionRepository.findByFilter(eq(BATCH_ID), eq(INITIATIVE_ID), anyList()))
-                .thenReturn(Flux.just(t1, t2));
-        when(rewardTransactionRepository.save(any())).thenAnswer(inv -> Mono.just(inv.getArgument(0)));
-
-        StepVerifier.create(service.updateAndSaveRewardTransactionsSuspended(BATCH_ID, INITIATIVE_ID, BATCH_ID_2, "2025-12"))
-                .expectNext(100L)
-                .verifyComplete();
-
-        assertEquals(BATCH_ID_2, t1.getRewardBatchId());
-        assertEquals("2025-12", t1.getRewardBatchLastMonthElaborated());
-        assertEquals(BATCH_ID_2, t2.getRewardBatchId());
-        assertEquals("2025-11", t2.getRewardBatchLastMonthElaborated());
-    }
-
-    @Test
-    void validateRewardBatch_notFound() {
-        when(rewardBatchRepository.findById(BATCH_ID)).thenReturn(Mono.empty());
-
-        StepVerifier.create(service.validateRewardBatch(OP1, INITIATIVE_ID, BATCH_ID))
-                .expectError(RewardBatchNotFound.class)
-                .verify();
-    }
-
-    @Test
-    void validateRewardBatch_L1_to_L2_success() {
-        RewardBatch b = RewardBatch.builder()
-                .id(BATCH_ID)
-                .assigneeLevel(RewardBatchAssignee.L1)
-                .numberOfTransactions(100L)
-                .numberOfTransactionsElaborated(20L)
-                .build();
-
-        when(rewardBatchRepository.findById(BATCH_ID)).thenReturn(Mono.just(b));
-        when(rewardBatchRepository.save(any())).thenAnswer(inv -> Mono.just(inv.getArgument(0)));
-
-        StepVerifier.create(service.validateRewardBatch(OP1, INITIATIVE_ID, BATCH_ID))
-                .assertNext(updated -> assertEquals(RewardBatchAssignee.L2, updated.getAssigneeLevel()))
-                .verifyComplete();
-    }
-
-    @Test
-    void validateRewardBatch_L1_wrongRole() {
-        RewardBatch b = RewardBatch.builder()
-                .id(BATCH_ID)
-                .assigneeLevel(RewardBatchAssignee.L1)
-                .numberOfTransactions(100L)
-                .numberOfTransactionsElaborated(20L)
-                .build();
-
-        when(rewardBatchRepository.findById(BATCH_ID)).thenReturn(Mono.just(b));
-
-        StepVerifier.create(service.validateRewardBatch("guest", INITIATIVE_ID, BATCH_ID))
-                .expectError(RoleNotAllowedForL1PromotionException.class)
-                .verify();
-    }
-
-    @Test
-    void validateRewardBatch_L1_not15percent() {
-        RewardBatch b = RewardBatch.builder()
-                .id(BATCH_ID)
-                .assigneeLevel(RewardBatchAssignee.L1)
-                .numberOfTransactions(100L)
-                .numberOfTransactionsElaborated(10L)
-                .build();
-
-        when(rewardBatchRepository.findById(BATCH_ID)).thenReturn(Mono.just(b));
-
-        StepVerifier.create(service.validateRewardBatch(OP1, INITIATIVE_ID, BATCH_ID))
-                .expectError(BatchNotElaborated15PercentException.class)
-                .verify();
-    }
-
-    @Test
-    void validateRewardBatch_L2_to_L3_success() {
-        RewardBatch b = RewardBatch.builder().id(BATCH_ID).assigneeLevel(RewardBatchAssignee.L2).build();
-        when(rewardBatchRepository.findById(BATCH_ID)).thenReturn(Mono.just(b));
-        when(rewardBatchRepository.save(any())).thenAnswer(inv -> Mono.just(inv.getArgument(0)));
-
-        StepVerifier.create(service.validateRewardBatch(OP2, INITIATIVE_ID, BATCH_ID))
-                .assertNext(updated -> assertEquals(RewardBatchAssignee.L3, updated.getAssigneeLevel()))
-                .verifyComplete();
-    }
-
-    @Test
-    void validateRewardBatch_L2_wrongRole() {
-        RewardBatch b = RewardBatch.builder().id(BATCH_ID).assigneeLevel(RewardBatchAssignee.L2).build();
-        when(rewardBatchRepository.findById(BATCH_ID)).thenReturn(Mono.just(b));
-
-        StepVerifier.create(service.validateRewardBatch("guest", INITIATIVE_ID, BATCH_ID))
-                .expectError(RoleNotAllowedForL2PromotionException.class)
-                .verify();
-    }
-
-    @Test
-    void validateRewardBatch_invalidState() {
-        RewardBatch b = RewardBatch.builder().id(BATCH_ID).assigneeLevel(RewardBatchAssignee.L3).build();
-        when(rewardBatchRepository.findById(BATCH_ID)).thenReturn(Mono.just(b));
-
-        StepVerifier.create(service.validateRewardBatch(OP3, INITIATIVE_ID, BATCH_ID))
-                .expectError(InvalidBatchStateForPromotionException.class)
-                .verify();
-    }
-
-    @Test
-    void generateAndSaveCsv_invalidBatchId_fastFail() {
-        StepVerifier.create(service.generateAndSaveCsv("bad/..", INITIATIVE_ID, MERCHANT_ID))
-                .expectError(IllegalArgumentException.class)
-                .verify();
-
-        verify(rewardBatchRepository, never()).findById(anyString());
-    }
-
-    @Test
-    void generateAndSaveCsv_success_withCFandWithoutCF() {
-        RewardBatch batch = RewardBatch.builder()
-                .id(BATCH_ID)
-                .merchantId(MERCHANT_ID)
-                .businessName("Biz")
-                .name("dicembre 2025")
-                .posType(PHYSICAL)
-                .build();
-
-        when(rewardBatchRepository.findById(BATCH_ID)).thenReturn(Mono.just(batch));
-        when(rewardBatchRepository.save(any())).thenAnswer(inv -> Mono.just(inv.getArgument(0)));
-
-        RewardTransaction trxWithCF = RewardTransaction.builder()
-                .id("T1")
-                .trxChargeDate(LocalDateTime.of(2025, 12, 10, 10, 30))
-                .fiscalCode("CF1")
-                .trxCode("CODE")
-                .effectiveAmountCents(1000L)
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.APPROVED)
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(100L).build()))
-                .additionalProperties(Map.of("productName", "Lavatrice", "productGtin", "803"))
-                .invoiceData(it.gov.pagopa.idpay.transactions.dto.InvoiceData.builder().docNumber("DOC").filename("inv.pdf").build())
-                .franchiseName("Store")
-                .build();
-
-        RewardTransaction trxNoCF = RewardTransaction.builder()
-                .id("T2")
-                .userId("U2")
-                .fiscalCode(null)
-                .trxCode("CODE2")
-                .effectiveAmountCents(2000L)
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.REJECTED)
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(50L).build()))
-                .additionalProperties(Map.of("productName", "Aspirapolvere", "productGtin", "123"))
-                .invoiceData(it.gov.pagopa.idpay.transactions.dto.InvoiceData.builder().docNumber(null).filename("inv2.pdf").build())
-                .franchiseName("Store2")
-                .build();
-
-        when(rewardTransactionRepository.findByFilter(eq(BATCH_ID), eq(INITIATIVE_ID), anyList()))
-                .thenReturn(Flux.just(trxWithCF, trxNoCF));
-
-        when(userRestClient.retrieveUserInfo("U2"))
-                .thenReturn(Mono.just(it.gov.pagopa.idpay.transactions.connector.rest.dto.UserInfoPDV.builder().pii("CF2").build()));
-
-        doReturn(Mono.just("some/path/file.csv"))
-                .when(serviceSpy).uploadCsvToBlob(anyString(), anyString());
-
-        StepVerifier.create(serviceSpy.generateAndSaveCsv(BATCH_ID, INITIATIVE_ID, MERCHANT_ID))
-                .assertNext(filename -> assertTrue(filename.endsWith(".csv")))
-                .verifyComplete();
-
-        verify(rewardBatchRepository).save(argThat(b -> b.getFilename() != null && b.getFilename().endsWith(".csv")));
-        assertEquals("CF2", trxNoCF.getFiscalCode());
-    }
-
-    @Test
-    void generateAndSaveCsv_uploadFails_propagates() {
-        RewardBatch batch = RewardBatch.builder()
-                .id(BATCH_ID)
-                .merchantId(MERCHANT_ID)
-                .businessName("Biz")
-                .name("dicembre 2025")
-                .posType(PHYSICAL)
-                .build();
-
-        when(rewardBatchRepository.findById(BATCH_ID)).thenReturn(Mono.just(batch));
-
-        RewardTransaction trx = RewardTransaction.builder()
-                .id("T1")
-                .fiscalCode("CF")
-                .trxCode("CODE")
-                .effectiveAmountCents(1000L)
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.APPROVED)
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(100L).build()))
-                .additionalProperties(Map.of())
-                .invoiceData(it.gov.pagopa.idpay.transactions.dto.InvoiceData.builder().filename("inv.pdf").build())
-                .franchiseName("Store")
-                .build();
-
-        when(rewardTransactionRepository.findByFilter(eq(BATCH_ID), eq(INITIATIVE_ID), anyList()))
-                .thenReturn(Flux.just(trx));
-
-        doReturn(Mono.error(new RuntimeException("upload fail")))
-                .when(serviceSpy).uploadCsvToBlob(anyString(), anyString());
-
-        StepVerifier.create(serviceSpy.generateAndSaveCsv(BATCH_ID, INITIATIVE_ID, MERCHANT_ID))
-                .expectError(RuntimeException.class)
-                .verify();
-
-        verify(rewardBatchRepository, never()).save(any());
-    }
-
-    @Test
-    void csvField_private_behavior() {
-        String r1 = ReflectionTestUtils.invokeMethod(service, "csvField", (String) null);
-        assertEquals("", r1);
-
-        String r2 = ReflectionTestUtils.invokeMethod(service, "csvField", "plain");
-        assertEquals("plain", r2);
-
-        String r3 = ReflectionTestUtils.invokeMethod(service, "csvField", "a;b");
-        assertEquals("\"a;b\"", r3);
-
-        String r4 = ReflectionTestUtils.invokeMethod(service, "csvField", "a\"b");
-        assertEquals("\"a\"\"b\"", r4);
-    }
-
-    @Test
-    void mapTransactionToCsvRow_private_handlesNullsAndFormatting() {
-        RewardTransaction trx = RewardTransaction.builder()
-                .id("T1")
-                .trxChargeDate(null)
-                .fiscalCode("CF")
-                .trxCode("CODE")
-                .effectiveAmountCents(null)
-                .rewardBatchTrxStatus(RewardBatchTrxStatus.APPROVED)
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(null).build()))
-                .additionalProperties(Map.of("productName", "Prod;X", "productGtin", "GTIN"))
-                .invoiceData(it.gov.pagopa.idpay.transactions.dto.InvoiceData.builder().docNumber(null).filename("inv.pdf").build())
-                .franchiseName("Store")
-                .build();
-
-        String row = ReflectionTestUtils.invokeMethod(service, "mapTransactionToCsvRow", trx, INITIATIVE_ID);
-        assertNotNull(row);
-        assertTrue(row.contains("T1"));
-        assertTrue(row.contains("\"Prod;X"));
-    }
-
-    @Test
-    void uploadCsvToBlob_success_status201() {
-        @SuppressWarnings("unchecked")
-        Response<BlockBlobItem> resp = Mockito.mock(Response.class);
-        when(resp.getStatusCode()).thenReturn(HttpStatus.CREATED.value());
-
-        when(approvedRewardBatchBlobService.upload(any(), anyString(), anyString()))
-                .thenReturn(resp);
-
-        StepVerifier.create(service.uploadCsvToBlob("file.csv", "content"))
-                .expectNext("file.csv")
-                .verifyComplete();
-    }
-
-    @Test
-    void uploadCsvToBlob_statusNot201_throwsClientExceptionWithBody() {
-        @SuppressWarnings("unchecked")
-        Response<BlockBlobItem> resp = Mockito.mock(Response.class);
-        when(resp.getStatusCode()).thenReturn(HttpStatus.BAD_REQUEST.value());
-
-        when(approvedRewardBatchBlobService.upload(any(), anyString(), anyString()))
-                .thenReturn(resp);
-
-        StepVerifier.create(service.uploadCsvToBlob("file.csv", "content"))
-                .expectErrorSatisfies(ex -> {
-                    assertInstanceOf(ClientExceptionWithBody.class, ex);
-                    ClientExceptionWithBody ce = (ClientExceptionWithBody) ex;
-                    assertEquals(ExceptionConstants.ExceptionCode.GENERIC_ERROR, ce.getCode());
-                })
-                .verify();
-    }
-
-
-    @Test
-    void postponeTransaction_transactionNotFound() {
-        when(rewardTransactionRepository.findTransactionInBatch(MERCHANT_ID, BATCH_ID, "T1"))
-                .thenReturn(Mono.empty());
-
-        StepVerifier.create(service.postponeTransaction(MERCHANT_ID, INITIATIVE_ID, BATCH_ID, "T1", LocalDate.now()))
-                .expectError(ClientExceptionNoBody.class)
-                .verify();
-    }
-
-    @Test
-    void postponeTransaction_batchNotFound() {
-        RewardTransaction trx = RewardTransaction.builder()
-                .id("T1")
-                .merchantId(MERCHANT_ID)
-                .rewardBatchId(BATCH_ID)
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(100L).build()))
-                .build();
-
-        when(rewardTransactionRepository.findTransactionInBatch(MERCHANT_ID, BATCH_ID, "T1"))
-                .thenReturn(Mono.just(trx));
-
-        when(rewardBatchRepository.findById(BATCH_ID)).thenReturn(Mono.empty());
-
-        StepVerifier.create(service.postponeTransaction(MERCHANT_ID, INITIATIVE_ID, BATCH_ID, "T1", LocalDate.now()))
-                .expectError(ClientExceptionWithBody.class)
-                .verify();
-    }
-
-    @Test
-    void postponeTransaction_invalidBatchStatus() {
-        RewardTransaction trx = RewardTransaction.builder()
-                .id("T1")
-                .merchantId(MERCHANT_ID)
-                .rewardBatchId(BATCH_ID)
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(100L).build()))
-                .build();
-
-        RewardBatch current = RewardBatch.builder()
-                .id(BATCH_ID)
-                .merchantId(MERCHANT_ID)
-                .businessName(BUSINESS_NAME)
-                .posType(PHYSICAL)
-                .month("2025-12")
-                .status(RewardBatchStatus.SENT)
-                .build();
-
-        when(rewardTransactionRepository.findTransactionInBatch(MERCHANT_ID, BATCH_ID, "T1"))
-                .thenReturn(Mono.just(trx));
-        when(rewardBatchRepository.findById(BATCH_ID)).thenReturn(Mono.just(current));
-
-        StepVerifier.create(service.postponeTransaction(MERCHANT_ID, INITIATIVE_ID, BATCH_ID, "T1", LocalDate.now()))
-                .expectError(ClientExceptionWithBody.class)
-                .verify();
-    }
-
-    @Test
-    void postponeTransaction_limitExceeded() {
-        RewardTransaction trx = RewardTransaction.builder()
-                .id("T1")
-                .merchantId(MERCHANT_ID)
-                .rewardBatchId(BATCH_ID)
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(100L).build()))
-                .build();
-
-        RewardBatch current = RewardBatch.builder()
-                .id(BATCH_ID)
-                .merchantId(MERCHANT_ID)
-                .businessName(BUSINESS_NAME)
-                .posType(PHYSICAL)
-                .month("2026-12")
-                .status(RewardBatchStatus.CREATED)
-                .build();
-
-        LocalDate initiativeEnd = LocalDate.of(2026, 1, 6);
-
-        when(rewardTransactionRepository.findTransactionInBatch(MERCHANT_ID, BATCH_ID, "T1"))
-                .thenReturn(Mono.just(trx));
-        when(rewardBatchRepository.findById(BATCH_ID)).thenReturn(Mono.just(current));
-
-        StepVerifier.create(service.postponeTransaction(MERCHANT_ID, INITIATIVE_ID, BATCH_ID, "T1", initiativeEnd))
-                .expectError(ClientExceptionWithBody.class)
-                .verify();
-    }
-
-    @Test
-    void postponeTransaction_success_movesAndUpdatesTrx() {
-        RewardTransaction trx = RewardTransaction.builder()
-                .id("T1")
-                .merchantId(MERCHANT_ID)
-                .rewardBatchId(BATCH_ID)
-                .rewards(Map.of(INITIATIVE_ID, Reward.builder().accruedRewardCents(100L).build()))
-                .build();
-
-        RewardBatch current = RewardBatch.builder()
-                .id(BATCH_ID)
-                .merchantId(MERCHANT_ID)
-                .businessName(BUSINESS_NAME)
-                .posType(PHYSICAL)
-                .month("2026-01")
-                .status(RewardBatchStatus.CREATED)
-                .build();
-
-        RewardBatch next = RewardBatch.builder()
-                .id(BATCH_ID_2)
-                .merchantId(MERCHANT_ID)
-                .businessName(BUSINESS_NAME)
-                .posType(PHYSICAL)
-                .month("2026-02")
-                .status(RewardBatchStatus.CREATED)
-                .build();
-
-        when(rewardTransactionRepository.findTransactionInBatch(MERCHANT_ID, BATCH_ID, "T1"))
-                .thenReturn(Mono.just(trx));
-        when(rewardBatchRepository.findById(BATCH_ID)).thenReturn(Mono.just(current));
-
-        doReturn(Mono.just(next)).when(serviceSpy).findOrCreateBatch(MERCHANT_ID, PHYSICAL, "2026-02", BUSINESS_NAME);
-
-        when(rewardTransactionRepository.save(any())).thenAnswer(inv -> Mono.just(inv.getArgument(0)));
-
-        when(rewardBatchRepository.updateTotals(eq(BATCH_ID), any()))
-                .thenReturn(Mono.empty());
-
-        when(rewardBatchRepository.updateTotals(eq(BATCH_ID_2), any()))
-                .thenReturn(Mono.empty());
-
-
-        StepVerifier.create(serviceSpy.postponeTransaction(MERCHANT_ID, INITIATIVE_ID, BATCH_ID, "T1", LocalDate.of(2026, 1, 6)))
-                .verifyComplete();
-
-        verify(rewardBatchRepository, times(2))
-                .updateTotals(anyString(), any());
-
-
-
-        assertEquals(BATCH_ID_2, trx.getRewardBatchId());
-        assertNotNull(trx.getRewardBatchInclusionDate());
-        assertNotNull(trx.getUpdateDate());
-    }
-
-    @Test
-    void deleteEmptyRewardBatches_deletesMatching() {
-        RewardBatch b1 = RewardBatch.builder().id("D1").month("2025-10").numberOfTransactions(0L).build();
-        RewardBatch b2 = RewardBatch.builder().id("D2").month("2025-09").numberOfTransactions(0L).build();
-
-        com.mongodb.reactivestreams.client.MongoDatabase db = mock(com.mongodb.reactivestreams.client.MongoDatabase.class);
-        when(db.getName()).thenReturn("db");
-
-        when(reactiveMongoTemplate.getMongoDatabase()).thenReturn(Mono.just(db));
-        when(reactiveMongoTemplate.getCollectionName(RewardBatch.class)).thenReturn("rewardBatch");
-
-        when(reactiveMongoTemplate.count(any(Query.class), eq(RewardBatch.class)))
-                .thenReturn(Mono.just(10L))
-                .thenReturn(Mono.just(2L));
-
-        when(reactiveMongoTemplate.find(any(Query.class), eq(RewardBatch.class)))
-                .thenReturn(Flux.just(b1, b2));
-
-        when(reactiveMongoTemplate.remove(any(Query.class), eq(RewardBatch.class)))
-                .thenReturn(Mono.just(DeleteResult.acknowledged(1L)));
-
-        StepVerifier.create(service.deleteEmptyRewardBatches())
-                .verifyComplete();
-
-        verify(reactiveMongoTemplate, times(2)).remove(any(Query.class), eq(RewardBatch.class));
-    }
-
-    @Test
-    void deleteEmptyRewardBatches_noMatches() {
-        com.mongodb.reactivestreams.client.MongoDatabase db = mock(com.mongodb.reactivestreams.client.MongoDatabase.class);
-        when(db.getName()).thenReturn("db");
-
-        when(reactiveMongoTemplate.getMongoDatabase()).thenReturn(Mono.just(db));
-        when(reactiveMongoTemplate.getCollectionName(RewardBatch.class)).thenReturn("rewardBatch");
-
-        when(reactiveMongoTemplate.count(any(Query.class), eq(RewardBatch.class)))
-                .thenReturn(Mono.just(10L))
-                .thenReturn(Mono.just(0L));
-
-        when(reactiveMongoTemplate.find(any(Query.class), eq(RewardBatch.class)))
-                .thenReturn(Flux.empty());
-
-        StepVerifier.create(service.deleteEmptyRewardBatches())
-                .verifyComplete();
-
-        verify(reactiveMongoTemplate, never()).remove(any(Query.class), eq(RewardBatch.class));
-    }
-
-    @Test
-    void postponeTransaction_whenTransactionIsSuspended_shouldUpdateSuspendedCounters() {
-
-        String merchantId = MERCHANT_ID;
-        String initiativeId = INITIATIVE_ID;
-        String rewardBatchId = BATCH_ID;
-        String transactionId = "TRX_ID";
-
-        LocalDate initiativeEndDate = LocalDate.of(2026, 12, 31);
-
-        long accruedRewardCents = 100L;
-
-        RewardTransaction trx = new RewardTransaction();
-        trx.setId(transactionId);
-        trx.setRewardBatchId(rewardBatchId);
-        trx.setRewardBatchTrxStatus(RewardBatchTrxStatus.SUSPENDED);
-
-        Reward reward = new Reward();
-        reward.setAccruedRewardCents(accruedRewardCents);
-
-        Map<String, Reward> rewardsMap = new HashMap<>();
-        rewardsMap.put(initiativeId, reward);
-        trx.setRewards(rewardsMap);
-
-        RewardBatch currentBatch = new RewardBatch();
-        currentBatch.setId(rewardBatchId);
-        currentBatch.setStatus(RewardBatchStatus.CREATED);
-        currentBatch.setMonth("2026-01");
-        currentBatch.setMerchantId(merchantId);
-        currentBatch.setBusinessName(BUSINESS_NAME);
-
-        RewardBatch nextBatch = new RewardBatch();
-        nextBatch.setId(BATCH_ID_2);
-        nextBatch.setStatus(RewardBatchStatus.CREATED);
-
-        when(rewardTransactionRepository.findTransactionInBatch(merchantId, rewardBatchId, transactionId))
-                .thenReturn(Mono.just(trx));
-
-        when(rewardBatchRepository.findById(rewardBatchId))
-                .thenReturn(Mono.just(currentBatch));
-
-        doReturn(Mono.just(nextBatch)).when(serviceSpy)
-                .findOrCreateBatch(any(), any(), any(), any());
-
-        when(rewardBatchRepository.updateTotals(eq(rewardBatchId), any()))
-                .thenReturn(Mono.empty());
-
-        when(rewardBatchRepository.updateTotals(eq(BATCH_ID_2), any()))
-                .thenReturn(Mono.empty());
-
-        when(rewardTransactionRepository.save(any()))
-                .thenReturn(Mono.just(trx));
-
-        StepVerifier.create(
-                        serviceSpy.postponeTransaction(
-                                merchantId,
-                                initiativeId,
-                                rewardBatchId,
-                                transactionId,
-                                initiativeEndDate
-                        )
-                )
-                .verifyComplete();
-
-        verify(rewardBatchRepository, times(2))
-                .updateTotals(anyString(), any(BatchCountersDTO.class));
-
-        verify(rewardTransactionRepository).save(trx);
-    }
-
-    @Test
-    void checkRewardBatchesOutcomes_withIds_success() {
-
-        RewardBatch batch1 = RewardBatch.builder().id(BATCH_ID).status(RewardBatchStatus.PENDING_REFUND).build();
-        RewardBatch batch2 = RewardBatch.builder().id(BATCH_ID_2).status(RewardBatchStatus.PENDING_REFUND).build();
-
-        ErogazioneOutcomeDTO erogazione1 = ErogazioneOutcomeDTO.builder()
-                .status("COMPLETATO")
-                .dateValue(LocalDate.now())
-                .build();
-
-        InvitaliaOutcomeResponseDTO outcome1 = InvitaliaOutcomeResponseDTO.builder()
-                .message(null)
-                .erogazione(erogazione1)
-                .build();
-
-        ErogazioneOutcomeDTO erogazione2 = ErogazioneOutcomeDTO.builder()
-                .status("RIFIUTATO")
-                .build();
-
-        ErrorDTO error = new ErrorDTO("ERR01", "Errore");
-
-        InvitaliaOutcomeResponseDTO outcome2 = InvitaliaOutcomeResponseDTO.builder()
-                .erogazione(erogazione2)
-                .errors(List.of(error))
-                .build();
-
-        when(rewardBatchRepository.findByIdAndStatus(BATCH_ID, RewardBatchStatus.PENDING_REFUND))
-                .thenReturn(Mono.just(batch1));
-        when(rewardBatchRepository.findByIdAndStatus(BATCH_ID_2, RewardBatchStatus.PENDING_REFUND))
-                .thenReturn(Mono.just(batch2));
-
-        when(erogazioniRestClient.getOutcome(BATCH_ID)).thenReturn(Mono.just(outcome1));
-        when(erogazioniRestClient.getOutcome(BATCH_ID_2)).thenReturn(Mono.just(outcome2));
-
-        when(rewardBatchRepository.save(any(RewardBatch.class)))
-                .thenAnswer(inv -> Mono.just(inv.getArgument(0)));
-
-        StepVerifier.create(service.checkRewardBatchesOutcomes(INITIATIVE_ID, List.of(BATCH_ID, BATCH_ID_2)))
-                .verifyComplete();
-
-        assertEquals(RewardBatchStatus.REFUNDED, batch1.getStatus());
-        assertNotNull(batch1.getRefundValutaDate());
-
-        assertEquals(RewardBatchStatus.NOT_REFUNDED, batch2.getStatus());
-        assertEquals("ERR01 - Errore", batch2.getRefundErrorMessage());
-
-        verify(rewardBatchRepository).findByIdAndStatus(BATCH_ID, RewardBatchStatus.PENDING_REFUND);
-        verify(rewardBatchRepository).findByIdAndStatus(BATCH_ID_2, RewardBatchStatus.PENDING_REFUND);
-        verify(erogazioniRestClient).getOutcome(BATCH_ID);
-        verify(erogazioniRestClient).getOutcome(BATCH_ID_2);
-        verify(rewardBatchRepository, times(2)).save(any());
-    }
-
-    @Test
-    void checkRewardBatchesOutcomes_emptyList_success() {
-        RewardBatch batch1 = RewardBatch.builder().id(BATCH_ID).status(RewardBatchStatus.PENDING_REFUND).build();
-
-        ErogazioneOutcomeDTO erogazione = ErogazioneOutcomeDTO.builder()
-                .status("COMPLETATO")
-                .dateValue(LocalDate.now())
-                .build();
-
-        InvitaliaOutcomeResponseDTO outcome = InvitaliaOutcomeResponseDTO.builder()
-                .message(null)
-                .erogazione(erogazione)
-                .build();
-
-        when(rewardBatchRepository.findByStatus(RewardBatchStatus.PENDING_REFUND))
-                .thenReturn(Flux.just(batch1));
-
-        when(erogazioniRestClient.getOutcome(BATCH_ID)).thenReturn(Mono.just(outcome));
-        when(rewardBatchRepository.save(any(RewardBatch.class))).thenAnswer(inv -> Mono.just(inv.getArgument(0)));
-
-        StepVerifier.create(service.checkRewardBatchesOutcomes(INITIATIVE_ID, null))
-                .verifyComplete();
-
-        assertEquals(RewardBatchStatus.REFUNDED, batch1.getStatus());
-        verify(rewardBatchRepository).findByStatus(RewardBatchStatus.PENDING_REFUND);
-        verify(erogazioniRestClient).getOutcome(BATCH_ID);
-        verify(rewardBatchRepository).save(batch1);
-    }
-
-    @Test
-    void updateBatch_completato_setsRefunded() {
-        RewardBatch batch = RewardBatch.builder().id(BATCH_ID).status(RewardBatchStatus.PENDING_REFUND).build();
-
-        ErogazioneOutcomeDTO erogazione = ErogazioneOutcomeDTO.builder()
-                .status("COMPLETATO")
-                .dateValue(LocalDate.now())
-                .build();
-
-        InvitaliaOutcomeResponseDTO outcome = InvitaliaOutcomeResponseDTO.builder()
-                .message(null)
-                .erogazione(erogazione)
-                .build();
-
-        when(rewardBatchRepository.save(any())).thenAnswer(inv -> Mono.just(inv.getArgument(0)));
-
-        StepVerifier.create(service.updateBatch(batch, outcome))
-                .assertNext(b -> {
-                    assertEquals(RewardBatchStatus.REFUNDED, b.getStatus());
-                    assertNotNull(b.getRefundValutaDate());
-                    assertNotNull(b.getRefundOutcomeTimestamp());
-                })
-                .verifyComplete();
-
-        verify(rewardBatchRepository).save(batch);
-    }
-
-    @Test
-    void updateBatch_rifiutato_setsNotRefunded_withoutErrors() {
-        RewardBatch batch = RewardBatch.builder()
-                .id(BATCH_ID)
-                .status(RewardBatchStatus.PENDING_REFUND)
-                .build();
-
-        ErogazioneOutcomeDTO erogazione = ErogazioneOutcomeDTO.builder()
-                .status("RIFIUTATO")
-                .build();
-
-        InvitaliaOutcomeResponseDTO outcome = InvitaliaOutcomeResponseDTO.builder()
-                .erogazione(erogazione)
-                .errors(null)
-                .build();
-
-        when(rewardBatchRepository.save(any())).thenAnswer(inv -> Mono.just(inv.getArgument(0)));
-
-        StepVerifier.create(service.updateBatch(batch, outcome))
-                .assertNext(b -> {
-                    assertEquals(RewardBatchStatus.NOT_REFUNDED, b.getStatus());
-                    assertNull(b.getRefundErrorMessage());
-                    assertNotNull(b.getRefundOutcomeTimestamp());
-                })
-                .verifyComplete();
-
-        verify(rewardBatchRepository).save(batch);
-    }
-
-    @Test
-    void updateBatch_inLavorazione_withCreatedStatus_setsPendingRefundTimestamp() {
-        RewardBatch batch = RewardBatch.builder()
-                .id(BATCH_ID)
-                .status(RewardBatchStatus.CREATED)
-                .build();
-
-        ErogazioneOutcomeDTO erogazione = ErogazioneOutcomeDTO.builder()
-                .status("IN_LAVORAZIONE")
-                .build();
-
-        InvitaliaOutcomeResponseDTO outcome = InvitaliaOutcomeResponseDTO.builder()
-                .erogazione(erogazione)
-                .build();
-
-        StepVerifier.create(service.updateBatch(batch, outcome))
-                .assertNext(b -> assertEquals(RewardBatchStatus.CREATED, b.getStatus()))
-                .verifyComplete();
-    }
-
-    @Test
-    void updateBatch_errore_withCreatedStatus_setsPendingRefundTimestamp() {
-        RewardBatch batch = RewardBatch.builder()
-                .id(BATCH_ID)
-                .status(RewardBatchStatus.CREATED)
-                .build();
-
-        ErogazioneOutcomeDTO erogazione = ErogazioneOutcomeDTO.builder()
-                .status("ERRORE")
-                .build();
-
-        InvitaliaOutcomeResponseDTO outcome = InvitaliaOutcomeResponseDTO.builder()
-                .erogazione(erogazione)
-                .build();
-
-        StepVerifier.create(service.updateBatch(batch, outcome))
-                .assertNext(b -> assertEquals(RewardBatchStatus.CREATED, b.getStatus()))
-                .verifyComplete();
-    }
-
-    @Test
-    void updateBatch_rifiutato_withoutErrors_setsNotRefunded() {
-
-        RewardBatch batch = RewardBatch.builder()
-                .id(BATCH_ID)
-                .status(RewardBatchStatus.PENDING_REFUND)
-                .build();
-
-        ErogazioneOutcomeDTO erogazione = ErogazioneOutcomeDTO.builder()
-                .status("RIFIUTATO")
-                .build();
-
-        InvitaliaOutcomeResponseDTO outcome = InvitaliaOutcomeResponseDTO.builder()
-                .erogazione(erogazione)
-                .errors(null)
-                .build();
-
-        when(rewardBatchRepository.save(any()))
-                .thenAnswer(inv -> Mono.just(inv.getArgument(0)));
-
-        StepVerifier.create(service.updateBatch(batch, outcome))
-                .assertNext(b -> {
-                    assertEquals(RewardBatchStatus.NOT_REFUNDED, b.getStatus());
-                    assertNull(b.getRefundErrorMessage());
-                })
-                .verifyComplete();
-    }
-
-    @Test
-    void updateBatch_rifiutato_withEmptyErrors_setsNotRefunded() {
-
-        RewardBatch batch = RewardBatch.builder()
-                .id(BATCH_ID)
-                .status(RewardBatchStatus.PENDING_REFUND)
-                .build();
-
-        ErogazioneOutcomeDTO erogazione = ErogazioneOutcomeDTO.builder()
-                .status("RIFIUTATO")
-                .build();
-
-        InvitaliaOutcomeResponseDTO outcome = InvitaliaOutcomeResponseDTO.builder()
-                .erogazione(erogazione)
-                .errors(List.of())
-                .build();
-
-        when(rewardBatchRepository.save(any()))
-                .thenAnswer(inv -> Mono.just(inv.getArgument(0)));
-
-        StepVerifier.create(service.updateBatch(batch, outcome))
-                .assertNext(b -> {
-                    assertEquals(RewardBatchStatus.NOT_REFUNDED, b.getStatus());
-                    assertNull(b.getRefundErrorMessage());
-                })
-                .verifyComplete();
-    }
-
-    @Test
-    void updateBatch_errore_withPendingStatus_keepsPendingRefund() {
-        RewardBatch batch = RewardBatch.builder()
-                .id(BATCH_ID)
-                .status(RewardBatchStatus.PENDING_REFUND)
-                .build();
-
-        ErogazioneOutcomeDTO erogazione = ErogazioneOutcomeDTO.builder()
-                .status("ERRORE")
-                .build();
-
-        InvitaliaOutcomeResponseDTO outcome = InvitaliaOutcomeResponseDTO.builder()
-                .erogazione(erogazione)
-                .build();
-
-        StepVerifier.create(service.updateBatch(batch, outcome))
-                .assertNext(b -> assertEquals(RewardBatchStatus.PENDING_REFUND, b.getStatus()))
-                .verifyComplete();
-    }
-
-    @Test
-    void updateBatch_inLavorazione_withPendingStatus_keepsPendingRefund() {
-        RewardBatch batch = RewardBatch.builder()
-                .id(BATCH_ID)
-                .status(RewardBatchStatus.PENDING_REFUND)
-                .build();
-
-        ErogazioneOutcomeDTO erogazione = ErogazioneOutcomeDTO.builder()
-                .status("IN_LAVORAZIONE")
-                .build();
-
-        InvitaliaOutcomeResponseDTO outcome = InvitaliaOutcomeResponseDTO.builder()
-                .erogazione(erogazione)
-                .build();
-
-        StepVerifier.create(service.updateBatch(batch, outcome))
-                .assertNext(b -> assertEquals(RewardBatchStatus.PENDING_REFUND, b.getStatus()))
-                .verifyComplete();
+        assertEquals("2026-01", RewardBatchSharedUtils.addOneMonth("2025-12"));
+        assertEquals("gennaio 2026", RewardBatchSharedUtils.addOneMonthToItalian("dicembre 2025"));
     }
 }
